requirements.txt
requests>=2.31
python-dotenv>=1.0
py-clob-client>=0.20
